change all float to double,

+++ driver.h
@@ -10,7 +10,7 @@ struct Driver
   virtual ~Driver() = default; 
 
 public:
-  virtual esphome::optional<std::map<std::string, float>> get_values(std::vector<unsigned char> &telegram) = 0;
+  virtual esphome::optional<std::map<std::string, double>> get_values(std::vector<unsigned char> &telegram) = 0;
   virtual esphome::optional<std::map<std::string, double>> get_values(std::vector<unsigned char> &telegram) = 0;
 
   std::string get_name() { return this->driver_type_; };
@@ -18,9 +18,9 @@ public:
 protected:
   Driver(std::string driver_type) : driver_type_(driver_type) {};
 
-  void add_to_map(std::map<std::string, float> &values,
+  void add_to_map(std::map<std::string, double> &values,
                   std::string name,
-                  esphome::optional<float> value) {
+                  esphome::optional<double> value) {
     if (value.has_value()) {
       values[name] = *value;
     }
@@ -43,8 +43,8 @@ protected:
   };
 
 // /
-  esphome::optional<float> get_0413(std::vector<unsigned char> &telegram) {
-    esphome::optional<float> ret_val{};
+  esphome::optional<double> get_0413(std::vector<unsigned char> &telegram) {
+    esphome::optional<double> ret_val{};
     uint32_t usage = 0;
     size_t i = 11;
     uint32_t total_register = 0x0413;
@@ -62,8 +62,8 @@ protected:
     return ret_val;
   };
 
-  esphome::optional<float> get_0C0E(std::vector<unsigned char> &telegram) {
-    esphome::optional<float> ret_val{};
+  esphome::optional<double> get_0C0E(std::vector<unsigned char> &telegram) {
+    esphome::optional<double> ret_val{};
     uint32_t usage = 0;
     size_t i = 11;
     uint32_t total_register = 0x0C0E;
@@ -81,8 +81,8 @@ protected:
     return ret_val;
   };
 
-  esphome::optional<float> get_0C0D(std::vector<unsigned char> &telegram) {
-    esphome::optional<float> ret_val{};
+  esphome::optional<double> get_0C0D(std::vector<unsigned char> &telegram) {
+    esphome::optional<double> ret_val{};
     uint32_t usage = 0;
     size_t i = 11;
     uint32_t total_register = 0x0C0D;
@@ -100,8 +100,8 @@ protected:
     return ret_val;
   };
 
-  esphome::optional<float> get_0C03(std::vector<unsigned char> &telegram) {
-    esphome::optional<float> ret_val{};
+  esphome::optional<double> get_0C03(std::vector<unsigned char> &telegram) {
+    esphome::optional<double> ret_val{};
     uint32_t usage = 0;
     size_t i = 11;
     uint32_t total_register = 0x0C03;
@@ -119,8 +119,8 @@ protected:
     return ret_val;
   };
 
-  esphome::optional<float> get_0C05(std::vector<unsigned char> &telegram) {
-    esphome::optional<float> ret_val{};
+  esphome::optional<double> get_0C05(std::vector<unsigned char> &telegram) {
+    esphome::optional<double> ret_val{};
     uint32_t usage = 0;
     size_t i = 11;
     uint32_t total_register = 0x0C05;
@@ -138,8 +138,8 @@ protected:
     return ret_val;
   };
 
-  esphome::optional<float> get_0C06(std::vector<unsigned char> &telegram) {
-    esphome::optional<float> ret_val{};
+  esphome::optional<double> get_0C06(std::vector<unsigned char> &telegram) {
+    esphome::optional<double> ret_val{};
     uint32_t usage = 0;
     size_t i = 11;
     uint32_t total_register = 0x0C06;
@@ -157,8 +157,8 @@ protected:
     return ret_val;
   };
 
-  esphome::optional<float> get_0C13(std::vector<unsigned char> &telegram) {
-    esphome::optional<float> ret_val{};
+  esphome::optional<double> get_0C13(std::vector<unsigned char> &telegram) {
+    esphome::optional<double> ret_val{};
     uint32_t usage = 0;
     size_t i = 11;
     uint32_t total_register = 0x0C13;
@@ -175,8 +175,8 @@ protected:
     return ret_val;
   };
 
-  esphome::optional<float> get_0E0A(std::vector<unsigned char> &telegram) {
-    esphome::optional<float> ret_val{};
+  esphome::optional<double> get_0E0A(std::vector<unsigned char> &telegram) {
+    esphome::optional<double> ret_val{};
     uint32_t usage = 0;
     size_t i = 11;
     uint32_t total_register = 0x0E0A;
@@ -194,8 +194,8 @@ protected:
     return ret_val;
   };
 
-  esphome::optional<float> get_0E01(std::vector<unsigned char> &telegram) {
-    esphome::optional<float> ret_val{};
+  esphome::optional<double> get_0E01(std::vector<unsigned char> &telegram) {
+    esphome::optional<double> ret_val{};
     uint32_t usage = 0;
     size_t i = 11;
     uint32_t total_register = 0x0E01;
@@ -213,8 +213,8 @@ protected:
     return ret_val;
   };
 
-  esphome::optional<float> get_0A2D(std::vector<unsigned char> &telegram) {
-    esphome::optional<float> ret_val{};
+  esphome::optional<double> get_0A2D(std::vector<unsigned char> &telegram) {
+    esphome::optional<double> ret_val{};
     uint32_t usage = 0;
     size_t i = 11;
     uint32_t total_register = 0x0A2D;
@@ -232,8 +232,8 @@ protected:
     return ret_val;
   };
 
-  esphome::optional<float> get_0A5A(std::vector<unsigned char> &telegram) {
-    esphome::optional<float> ret_val{};
+  esphome::optional<double> get_0A5A(std::vector<unsigned char> &telegram) {
+    esphome::optional<double> ret_val{};
     uint32_t usage = 0;
     size_t i = 11;
     uint32_t total_register = 0x0A5A;
@@ -251,8 +251,8 @@ protected:
     return ret_val;
   };
 
-  esphome::optional<float> get_0A5E(std::vector<unsigned char> &telegram) {
-    esphome::optional<float> ret_val{};
+  esphome::optional<double> get_0A5E(std::vector<unsigned char> &telegram) {
+    esphome::optional<double> ret_val{};
     uint32_t usage = 0;
     size_t i = 11;
     uint32_t total_register = 0x0A5E;
@@ -270,8 +270,8 @@ protected:
     return ret_val;
   };
 
-  esphome::optional<float> get_0C943A(std::vector<unsigned char> &telegram) {
-    esphome::optional<float> ret_val{};
+  esphome::optional<double> get_0C943A(std::vector<unsigned char> &telegram) {
+    esphome::optional<double> ret_val{};
     uint32_t usage = 0;
     size_t i = 11;
     uint32_t total_register = 0x0C943A;
